Add firmware change support for Telit LE910C4
This adds a custom MMIfaceModem3gpp register_in_network implementation
that checks the firmware in use before registering with a network. If
the SIM is a Verizon SIM and the Verizon firmware is not in use, the
new register_in_network implementation switches the firmware to the
Verizon firmware. Similarly, if a non-Verizon SIM is in use along with
the Verizon firmware, then the firmware is switched to the non-Verizon
firmware before attempting to register with the network.